requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.2
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.5.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
ecos==2.0.14
exceptiongroup==1.3.0
executing==2.2.0
fastcluster==1.3.0
fastjsonschema==2.21.1
feather-format==0.4.1
filelock==3.18.0
fonttools==4.58.4
formulaic==1.2.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
harmony-pytorch==0.1.8
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
imagecodecs==2025.3.30
imbalanced-learn==0.13.0
imblearn==0.0
inflate64==1.0.3
interface-meta==1.3.0
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lifelines==0.30.0
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.6.3
multivolumefile==0.2.3
narwhals==1.48.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.2
numexpr==2.11.0
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openslide-bin==4.0.0.8
openslide-python==1.4.2
optuna==4.4.0
osqp==0.6.7.post3
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==1.0.0
pyarrow==20.0.0
pybcj==1.0.6
pycox==0.3.0
pycparser==2.22
pycryptodomex==3.23.0
Pygments==2.19.2
pyparsing==3.2.3
pyppmd==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
pyzstd==0.17.0
qdldl==0.1.7.post5
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scikit-survival==0.24.1
scipy==1.15.3
Send2Trash==1.8.3
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.16
tinycss2==1.4.0
tomli==2.2.1
torch==2.7.1
torch-geometric==2.6.1
torchinfo==1.8.0
torchtuples==0.2.2
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.2
yarl==1.20.1
